simplify test case
